+++ include/osqp/.gitignore
