+++ include/osqp/error.h
@@ -7,7 +7,7 @@
 extern "C" {
 # endif // ifdef __cplusplus
 
-# include "types.h"
+# include "osqp/types.h"
 
 
 /* OSQP error macro */
