+++ include/osqp/lin_alg.h
@@ -6,7 +6,7 @@
 extern "C" {
 # endif // ifdef __cplusplus
 
-# include "types.h"
+# include "osqp/types.h"
 
 
 /* VECTOR FUNCTIONS ----------------------------------------------------------*/
