+++ include/osqp/lin_sys.h
@@ -7,7 +7,7 @@
 extern "C" {
 # endif // ifdef __cplusplus
 
-# include "types.h"
+# include "osqp/types.h"
 
 /**
  * Load linear system solver shared library
