+++ include/osqp/polish.h
@@ -7,7 +7,7 @@ extern "C" {
 # endif // ifdef __cplusplus
 
 
-# include "types.h"
+# include "osqp/types.h"
 
 /**
  * Solution polish: Solve equality constrained QP with assumed active
